+++ plugin_info/configuration.php
@@ -29,6 +29,13 @@ if (!isConnect('admin')) {
      <input class="configKey form-control" data-l1key="cronMinute" disabled/>
    </div>
 	</div>
+    <div class="form-group">
+			<label class="col-lg-4 control-label" title="{{Suivant les régions, GRDF donne la dernière consommation journaliere à J-1, J-2... Pour éviter au plugin d'intérroger le site pour rien, indiquez le nombre de jours de retard généralement constaté}}">{{Nombre de jours de retard}}</label>
+			<div class="col-lg-4">
+			<input class="configKey form-control" data-l1key="days-delay"/>	
+			</div>
+		</div>
+		</div>
    <div class="form-group">
 			<label class="col-lg-4 control-label" title="{{Ajouter un message dans le centre de message lorsqu'un captcha est détecté}}">{{Warning captcha}}</label>
 			<div class="col-lg-4">