+++ docs/en_US/changelog.md
@@ -4,9 +4,10 @@
 >
 >As a reminder if there is no information on the update, it means that it only concerns the updating of documentation, translation or text.
 
-# Beta - 26/01/2022
+# Beta - 01/02/2022
 - Fix for absent comparison data
 - New plugin configuration options to react when a captcha is detected
+- New plugin configuration to specify number of days delay 
 
 # Stable - 24/01/2022
 - Fix on monthly consumption computation for the last year month